requirements.txt
argon2-cffi
CTkMessagebox
customtkinter
pycryptodome
pyotp
pyperclip
Requests
setuptools
